requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiohttp-client-cache==0.6.1
aiosignal==1.3.1
appdirs==1.4.4
astroid==2.8.6
async-timeout==4.0.2
attrs==21.4.0
Brotli==1.0.9
cattrs==22.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
cycler==0.11.0
exceptiongroup==1.1.0
fonttools==4.33.2
frozenlist==1.3.3
idna==3.3
isort==5.10.1
itsdangerous==2.1.2
kiwisolver==1.4.2
lazy-object-proxy==1.6.0
line-profiler==4.0.2
lxml==4.8.0
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
multidict==6.0.4
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pdf2image==1.16.2
Pillow==9.1.0
platformdirs==2.4.0
plumbum==1.7.2
pycares==4.3.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
pylint==2.11.1
PyNaCl==1.5.0
pyparsing==3.0.6
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.0
python-forge==18.6.0
pytz==2022.1
requests==2.28.2
requests-cache==0.9.8
simplejson==3.18.1
six==1.16.0
sortedcontainers==2.4.0
toml==0.10.2
typing_extensions==4.4.0
url-normalize==1.4.3
urllib3==1.26.8
wrapt==1.13.3
yarl==1.8.2
